server/requirements.txt
asgiref==3.7.2
distlib==0.3.8
Django==4.2.10
django-cors-headers==4.3.1
djangorestframework==3.14.0
drf-yasg==1.21.7
filelock==3.13.1
inflection==0.5.1
packaging==23.2
platformdirs==4.2.0
pytz==2024.1
PyYAML==6.0.1
sqlparse==0.4.4
typing-extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
virtualenv==20.25.1
virtualenvwrapper-win==1.2.7
